requirements.txt
asgiref==2.3.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.1
Automat==0.7.0
cffi==1.12.3
channels==2.1.2
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.4
Flask==1.1.1
hyperlink==19.0.0
idna==2.8
imutils==0.5.2
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.0
opencv-contrib-python==4.1.0.25
pandas==0.25.0
Pillow==6.1.0
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.0
virtualenv==16.7.2
Werkzeug==0.15.5
wrapt==1.11.2
zmq==0.0.0
zope.interface==4.6.0
Whitenoise==4.1.3